refactor(scope): rename package to authscope
